# openapi.yaml
openapi: 3.0.0
info:
  title: Indikara Backend API
  version: 1.0.0
  description: API for authentication, products, and health checks
servers:
  - url: https://backend-wei5.onrender.com
    description: Deployment server
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/payu/webhook:
    post:
      summary: PayU webhook endpoint (receives payment notifications)
      tags:
        - PayU
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                key:
                  type: string
                txnid:
                  type: string
                amount:
                  type: string
                productinfo:
                  type: string
                firstname:
                  type: string
                email:
                  type: string
                status:
                  type: string
                hash:
                  type: string
            examples:
              success-demo:
                value:
                  key: your_key
                  txnid: tx_12345
                  amount: '100'
                  productinfo: Demo
                  firstname: Demo
                  email: demo@example.com
                  status: success
                  hash: ...
      responses:
        '200':
          description: OK
        '400':
          description: Invalid hash or payment
        '403':
          description: Forbidden (IP not allowed)
  /api/payu/initiate:
    post:
      summary: Initiate PayU hosted checkout for a pending order (server computes amount)
      tags:
        - PayU
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
            examples:
              demo:
                value:
                  orderId: order_12345
      responses:
        '200':
          description: Payment form data and payment URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  formData:
                    type: object
                  paymentUrl:
                    type: string
              examples:
                demo:
                  value:
                    formData:
                      key: your_key
                      txnid: tx_12345
                      amount: '100'
                    paymentUrl: https://secure.payu.in/_payment
...